move-core-types: move tests closer to implementation
